requirements.txt
python-dotenv>=1.0.1

# Pinecone + LangChain
pinecone>=5.4.2
langchain>=0.3.18
langchain-core>=0.3.34
langchain-community>=0.3.17
langchain-text-splitters>=0.3.6
langchain-pinecone>=0.2.2

# Document loading
pypdf>=5.3.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# UI (optional)
streamlit>=1.42.0
requests>=2.32.0

# MongoDB + LangGraph
pymongo>=4.7.0
langgraph>=0.2.0

# Gemini
google-genai>=0.6.0
